Separated out JSON serialization to Java Object
